Move useFlag to form-builder and use local action
