--- app/(gcforms)/[locale]/(form administration)/form-builder/[id]/actions.ts
@@ -1,6 +1,0 @@
-"use server";
-import { checkOne } from "@lib/cache/flags";
-
-export async function checkFlag(id: string) {
-  return checkOne(id);
-}
