+++ app/layout.tsx
@@ -10,7 +10,7 @@ import { Noto_Sans, Lato } from "next/font/google";
 import { googleTagManager } from "@lib/cspScripts";
 import { headers } from "next/headers";
 import { auth } from "@lib/auth";
-import { checkFlag } from "@formBuilder/[id]/actions";
+import { checkFlag } from "@formBuilder/actions";
 
 const notoSans = Noto_Sans({
   weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"],
